+++ src/main/java/nextstep/courses/domain/session/SessionPeriod.java
@@ -1,0 +1,18 @@
+package nextstep.courses.domain.session;
+
+import java.time.LocalDateTime;
+
+public class SessionPeriod {
+
+    private LocalDateTime startDate;
+
+    private LocalDateTime endDate;
+
+    public SessionPeriod() {
+    }
+
+    public SessionPeriod(LocalDateTime startDate, LocalDateTime endDate) {
+        this.startDate = startDate;
+        this.endDate = endDate;
+    }
+}
