+++ src/main/java/nextstep/courses/domain/session/Sessions.java
@@ -1,0 +1,16 @@
+package nextstep.courses.domain.session;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Sessions {
+    private final List<Session> value;
+
+    public Sessions() {
+        this.value = new ArrayList<>();
+    }
+
+    public boolean add(Session session) {
+        return value.add(session);
+    }
+}
