+++ src/main/java/nextstep/courses/domain/session/image/SessionImageType.java
@@ -1,0 +1,5 @@
+package nextstep.courses.domain.session.image;
+
+public enum SessionImageType {
+    GIF, JPG, JPEG, PNG, SVG
+}
