+++ src/main/java/nextstep/courses/domain/session/property/SessionStatus.java
@@ -1,0 +1,13 @@
+package nextstep.courses.domain.session.property;
+
+public enum SessionStatus {
+    PREPARING, ENROLLING, CLOSED;
+
+    public boolean canEnroll() {
+        return this == ENROLLING;
+    }
+
+    public boolean canNotEnroll() {
+        return !canEnroll();
+    }
+}
