+++ src/main/java/nextstep/courses/domain/session/property/SessionTypeEnrollStrategy.java
@@ -1,0 +1,7 @@
+package nextstep.courses.domain.session.property;
+
+import nextstep.courses.domain.session.constraint.SessionConstraint;
+
+public interface SessionTypeEnrollStrategy {
+    boolean canEnroll(SessionConstraint sessionConstraint, int enrollmentCount, long amount);
+}
