--- src/main/java/nextstep/session/domain/Sessions.java
@@ -1,16 +1,0 @@
-package nextstep.session.domain;
-
-import java.util.ArrayList;
-import java.util.List;
-
-public class Sessions {
-    private final List<Session> value;
-
-    public Sessions() {
-        this.value = new ArrayList<>();
-    }
-
-    public boolean add(Session session) {
-        return value.add(session);
-    }
-}
