--- src/main/java/nextstep/session/domain/constraint/SessionConstraint.java
@@ -1,25 +1,0 @@
-package nextstep.session.domain.constraint;
-
-public class SessionConstraint {
-
-    private final SessionFee fee;
-
-    private final SessionCapacity capacity;
-
-    public SessionConstraint(long fee, int capacity) {
-        this(new SessionFee(fee), new SessionCapacity(capacity));
-    }
-
-    public SessionConstraint(SessionFee fee, SessionCapacity capacity) {
-        this.fee = fee;
-        this.capacity = capacity;
-    }
-
-    public boolean isSameFee(long amount) {
-        return fee.isSame(amount);
-    }
-
-    public boolean isGreaterThenCapacity(int value) {
-        return capacity.isGreaterThan(value);
-    }
-}
