--- src/main/java/nextstep/session/domain/image/SessionImageType.java
@@ -1,5 +1,0 @@
-package nextstep.session.domain.image;
-
-public enum SessionImageType {
-    GIF, JPG, JPEG, PNG, SVG
-}
