--- src/main/java/nextstep/session/domain/property/SessionStatus.java
@@ -1,13 +1,0 @@
-package nextstep.session.domain.property;
-
-public enum SessionStatus {
-    PREPARING, ENROLLING, CLOSED;
-
-    public boolean canEnroll() {
-        return this == ENROLLING;
-    }
-
-    public boolean canNotEnroll() {
-        return !canEnroll();
-    }
-}
