--- src/main/java/nextstep/session/domain/property/SessionType.java
@@ -1,24 +1,0 @@
-package nextstep.session.domain.property;
-
-import nextstep.session.domain.constraint.SessionConstraint;
-
-public enum SessionType {
-    FREE((sessionConstraint, payments, payment) -> true),
-    PAID((sessionConstraint, enrollmentCount, amount) ->
-        sessionConstraint.isGreaterThenCapacity(enrollmentCount) && sessionConstraint.isSameFee(amount)
-    );
-
-    private final SessionTypeEnrollStrategy sessionTypeEnrollStrategy;
-
-    SessionType(SessionTypeEnrollStrategy sessionTypeEnrollStrategy) {
-        this.sessionTypeEnrollStrategy = sessionTypeEnrollStrategy;
-    }
-
-    public boolean canEnroll(SessionConstraint sessionConstraint, int enrollmentCount, long amount) {
-        return sessionTypeEnrollStrategy.canEnroll(sessionConstraint, enrollmentCount, amount);
-    }
-
-    public boolean canNotEnroll(SessionConstraint sessionConstraint, int enrollmentCount, long amount) {
-        return !canEnroll(sessionConstraint, enrollmentCount, amount);
-    }
-}
