--- src/main/java/nextstep/session/domain/property/SessionTypeEnrollStrategy.java
@@ -1,7 +1,0 @@
-package nextstep.session.domain.property;
-
-import nextstep.session.domain.constraint.SessionConstraint;
-
-public interface SessionTypeEnrollStrategy {
-    boolean canEnroll(SessionConstraint sessionConstraint, int enrollmentCount, long amount);
-}
