+++ src/test/java/nextstep/courses/domain/session/SessionsTest.java
@@ -1,0 +1,15 @@
+package nextstep.courses.domain.session;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+
+class SessionsTest {
+
+    @DisplayName("Sessions 인스턴스 생성")
+    @Test
+    public void testConstructor() {
+        assertDoesNotThrow(Sessions::new);
+    }
+}
